+++ bonus/srcs_bonus/render_bonus/shape_bonus/plane_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   plane_bonus.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: star <star@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: nrolland <nrolland@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/30 15:32:58 by star              #+#    #+#             */
-/*   Updated: 2025/09/30 17:25:33 by star             ###   ########.fr       */
+/*   Updated: 2025/10/01 19:42:21 by nrolland         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ float	hit_plane(t_vec c, t_vec normal, t_ray r)
 	float	denominator;
 
 	denominator = scalar_product(r.direction, normal);
-	if (fabs(denominator) < 1e-6)
+	if (fabsf(denominator) < 1e-6)
 		return (-1.0);
 	t = scalar_product(vec_sub(c, r.origin), normal) / denominator;
 	if (t >= 0.0)
