+++ manda/includes/display.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   display.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: star <star@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: nrolland <nrolland@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/28 15:48:29 by star              #+#    #+#             */
-/*   Updated: 2025/09/25 17:49:10 by star             ###   ########.fr       */
+/*   Updated: 2025/10/01 19:21:14 by nrolland         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@
 void	handle_low_resolution(t_data *data, int x, int y, int resolution);
 void	display_screen(t_data *data);
 void	sampling(t_fcolor *o_pixel, t_fcolor n_pixel,
-			double coef_new_p, double coef_old_p);
+			float coef_new_p, float coef_old_p);
 
 #endif
